actually run tests in travis
Need to look over these changes and document them, it's a bit of
a house of cards
but this appears to actually run the mocha tests in a docker environment.

tests running from previous commit would crash because can't test afaict
when using a meteor 'bundle' running with 'node main.js'.

need to rename tests.js because amusingly that doesnt' get picked up
as a filename for containing tests.