remove extra client-side Blobs2 collection declare
Package blobs, commit 814bd1e , creates the client-side Blobs2 collection. This was submodule'd into patient care develop branch but not master branch. However, medbook_ was updated with this commit of blobs when 'new commits' were pulled into it, not realizing it was not actually live yet. This extra creation of the blobs collection causes the following exception: "Uncaught Error: There is already a collection named "blobs"" and interferes with other modules being loaded (noticeably, autoforms.)