Add node_modules to git and exclude from docker
Previously when we run with dev-start.sh, both jobrunner & patient-
care complain of missing npm modules. Example from jobrunner:

Unable to resolve some modules:

  "q" in /app-dev/server/lib/globals.js (os.linux.x86_64)
  "temp" in /app-dev/server/lib/globals.js (os.linux.x86_64)
  "byline" in /app-dev/server/lib/globals.js (os.linux.x86_64)
  "mysql" in /app-dev/server/classes/UpdateCbioSecurity.js (os.linux.x86_64)

If you notice problems related to these missing modules, consider running:

  meteor npm install --save q temp byline mysql

When the docker containers are built for prod mode,
'meteor npm install' is run; it crashes if there is a node_modules folder
already present due to EXDEV errors.

However, when running in dev mode, I would have to ssh into the container
after it spun up and run 'meteor npm install'.

Thus, we add the node_modules folder to git so that they will be present
and dev mode will spin up ok; but we also add to the .dockerignore in root
that these folders are NOT to be copied over when we add the app dir to the
container.

Since the file changes are voluminous, here's a summary:

updated .dockerignore
removed the .gitignore in job-runner that contained only: node_modules

Added the job-runner node_modules folder to git
Added the patient-care node_modules folder to git