update job-runner to 1.1.3 for node_modules change
Nothing should be different about the actual built docker image, but since there's a code change let's play it safe.
Previously, job runner's node_modules folder was in .gitignore.
That has been removed & the modules have been added to git. In replacement, the root's .dockerignore is now ignoring that node_module directory so it is not copied into the built image.