Add default value type to standalone limma
Fixes 2 things:

-- right now the only possible value is gene_expression. It's annoying to have
to select gene expression every time when we know that's what the user wants.

-- the following bug @mokolodi1 mentioned in september is still active; this fix appears
to bypass it, empirically: i
"select a value type to create a limma job, then navigate away, then
go back and select a value type again. The value_type field’s
reactivity is broken and the other fields don’t show up."

Note :
this will error if there are ever 0 allowed values for what type of limma
to run.
However, if we are ever in that situation, we can't run limma anyhow, so it
does not matter.